feat: call api reading history
